[media] work around on a very strange bug!!!
on Buildroot:
if the calloc allocate (length + 1) as necessary,
for an absolute path the next call to calloc will crash.
for a relative path, the application runs fine.
if the calloc allocate (length + 2)
the application runs fine in any cases.
For the HOME with an absolute path (see under),
the application runs fine in any cases.